week2/pyproject.toml
[project]
name = "week2-rag"
version = "0.1.0"
description = "Supreme Court Precedent Finder — RAG with LangChain + Claude"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-chroma>=0.2.0",
    "langchain-huggingface>=0.1.0",
    "langchain-community>=0.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "jq>=1.11.0",
    "sentence-transformers>=5.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.11.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "SIM"]

[tool.ruff.lint.per-file-ignores]
"rag.py" = ["S101", "S113"]  # assert in eval, requests without timeout handled explicitly
"tests/**" = ["S101"]  # assert is fine in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow — real embeddings or LLM calls (deselect with '-m \"not slow\"')",
    "requires_api_key: marks tests that need ANTHROPIC_API_KEY",
]
